hoàn chỉnh

+++ micromouse1/Core/Inc/encoder.h
@@ -4,7 +4,7 @@
 #include "stm32f4xx_hal.h"
 
 void Encoder_Init(TIM_HandleTypeDef *htim_left, TIM_HandleTypeDef *htim_right);
-void Encoder_GetCounters(int16_t *cnt_left, int16_t *cnt_right);
+void Encoder_GetCounters(int32_t *cnt_left, int32_t *cnt_right);
 void Encoder_Reset(void);
 
 // Bo sung hàm doc quãng duong theo mm
